[interoperator] Move mcd provisioner to subfolder
