+++ interoperator/pkg/controller/multiclusterdeploy/add_provisoner.go
@@ -14,13 +14,13 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controller
+package multiclusterdeploy
 
 import (
-	"github.com/cloudfoundry-incubator/service-fabrik-broker/interoperator/pkg/controller/multiclusterdeploy/sfserviceinstancereplicator"
+	"github.com/cloudfoundry-incubator/service-fabrik-broker/interoperator/pkg/controller/multiclusterdeploy/provisioner"
 )
 
 func init() {
 	// AddToManagerFuncs is a list of functions to create controllers and add them to a manager.
-	AddToManagerFuncs = append(AddToManagerFuncs, sfserviceinstancereplicator.Add)
+	AddToManagerFuncs = append(AddToManagerFuncs, provisioner.Add)
 }
