+++ interoperator/pkg/controller/multiclusterdeploy/add_sfserviceinstancereplicator.go
@@ -1,5 +1,3 @@
-// +build multiclusterdeploy
-
 /*
 Copyright 2018 The Service Fabrik Authors.
 
@@ -16,13 +14,13 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controller
+package multiclusterdeploy
 
 import (
-	"github.com/cloudfoundry-incubator/service-fabrik-broker/interoperator/pkg/controller/multiclusterdeploy"
+	"github.com/cloudfoundry-incubator/service-fabrik-broker/interoperator/pkg/controller/multiclusterdeploy/sfserviceinstancereplicator"
 )
 
 func init() {
 	// AddToManagerFuncs is a list of functions to create controllers and add them to a manager.
-	AddToManagerFuncs = append(AddToManagerFuncs, multiclusterdeploy.Add)
+	AddToManagerFuncs = append(AddToManagerFuncs, sfserviceinstancereplicator.Add)
 }
