+++ interoperator/pkg/controller/multiclusterdeploy/multiclusterdeploy.go
@@ -1,0 +1,8 @@
+package multiclusterdeploy
+
+import (
+	"sigs.k8s.io/controller-runtime/pkg/manager"
+)
+
+// AddToManagerFuncs is a list of functions to add all Controllers to the Manager
+var AddToManagerFuncs []func(manager.Manager) error
